Fixed score display after loss. Reset Samuel's dimensions to fixed numbers. Minor code corrections.

+++ src/com/example/savingsamuel/GameStateManager.java
@@ -121,7 +121,7 @@ public class GameStateManager {
 	}
 	private void update() {
         long nuptime = SystemClock.uptimeMillis();
-        float elapsed = ((float) (nuptime - lUptime) / 1000f) * fTimescale;
+        float elapsed = ((nuptime - lUptime) / 1000f) * fTimescale;
         lUptime = nuptime;
         
         Timer.Update(elapsed);
